Extract editor sandbox into own component
